feat: upsert for update
