+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cypress-mongodb",
-  "version": "5.1.4",
+  "version": "5.2.1",
   "description": "Cypress MongoDB plugin",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
@@ -32,7 +32,7 @@
   "devDependencies": {
     "@faker-js/faker": "^7.5.0",
     "@types/node": "^18.11.9",
-    "cypress": "^12.3.0",
+    "cypress": "12.7.0",
     "prettier": "^2.8.0",
     "typescript": "^4.9.3"
   }
